Moved DSL part to own codegen
